+++ changelog.md
@@ -17,7 +17,7 @@
 * Include short-cuts to default repositories when creating workflow templates using the command-line interface.
 
 
-### 0.2.0 - 2020-08-??
+### 0.2.0 - 2020-07-30
 
 * Use SQLAlchemy as ORM for the database model.
 * Run workflow templates from command-line for test purposes.